+++ ios/Classes/FlutterArkit.m
@@ -157,6 +157,19 @@
     
 }
 
+- (void)renderer:(id <SCNSceneRenderer>)renderer didAddNode:(SCNNode *)node forAnchor:(ARAnchor *)anchor {
+    if (node.name == nil) {
+        node.name = [NSUUID UUID].UUIDString;
+    }
+    NSDictionary* params = [self prepareParamsForAnchorEventwithNode:node andAnchor:anchor];
+    [_channel invokeMethod: @"didAddNodeForAnchor" arguments: params];
+}
+
+- (void)renderer:(id <SCNSceneRenderer>)renderer didUpdateNode:(SCNNode *)node forAnchor:(ARAnchor *)anchor {
+    NSDictionary* params = [self prepareParamsForAnchorEventwithNode:node andAnchor:anchor];
+    [_channel invokeMethod: @"didUpdateNodeForAnchor" arguments: params];
+}
+
 #pragma mark - Lazy loads
 
 -(ARWorldTrackingConfiguration *)configuration {
@@ -345,5 +358,32 @@
     return debugOptions;
 }
 
+- (NSString*) convertSimdFloat4x4ToString: (simd_float4x4) matrix {
+    NSMutableString* ret = [NSMutableString stringWithCapacity:0];
+    for (int i = 0; i< 4; i++) {
+        for (int j = 0; j< 4; j++) {
+            [ret appendString:[NSString stringWithFormat:@"%f ", matrix.columns[i][j]]];
+        }
+    }
+    return ret;
+}
+
+- (NSString*) convertSimdFloat3ToString: (simd_float3) vector {
+    return [NSString stringWithFormat:@"%f %f %f", vector[0], vector[1], vector[2]];
+}
+
+- (NSDictionary<NSString*, NSString*>*) prepareParamsForAnchorEventwithNode: (SCNNode*) node andAnchor: (ARAnchor*) anchor {
+    NSMutableDictionary<NSString*, NSString*>* params = [@{@"node_name": node.name,
+                                                           @"identifier": [anchor.identifier UUIDString],
+                                                           @"transform": [self convertSimdFloat4x4ToString:anchor.transform]
+                                                           } mutableCopy];
+    if ([anchor isMemberOfClass:[ARPlaneAnchor class]]) {
+        ARPlaneAnchor *plane = (ARPlaneAnchor*)anchor;
+        [params setObject:@"planeAnchor" forKey:@"anchorType"];
+        [params setObject:[self convertSimdFloat3ToString:plane.center] forKey:@"center"];
+        [params setObject:[self convertSimdFloat3ToString:plane.extent] forKey:@"extent"];
+    }
+    return params;
+}
 
 @end