+++ lib/arkit_plugin.dart
@@ -1,5 +1,6 @@
 export 'package:arkit_plugin/widget/arkit_scene_view.dart';
 export 'package:arkit_plugin/widget/arkit_arplane_detection.dart';
+export 'package:arkit_plugin/geometries/arkit_anchor.dart';
 export 'package:arkit_plugin/geometries/arkit_geometry.dart';
 export 'package:arkit_plugin/geometries/arkit_material.dart';
 export 'package:arkit_plugin/geometries/arkit_vector3.dart';