+++ lib/geometries/arkit_vector3.dart
@@ -9,4 +9,9 @@ class ARKitVector3 {
   final double z;
 
   Map<String, double> toMap() => {'x': x, 'y': y, 'z': z};
+
+  static ARKitVector3 fromString(String str) {
+    final coords = str.split(' ').map((s) => double.parse(s)).toList();
+    return ARKitVector3(coords[0], coords[1], coords[2]);
+  }
 }