+++ lib/geometries/arkit_text.dart
@@ -1,24 +1,15 @@
 import 'package:arkit_plugin/geometries/arkit_geometry.dart';
 import 'package:arkit_plugin/geometries/arkit_material.dart';
 import 'package:meta/meta.dart';
-import 'package:vector_math/vector_math_64.dart';
 
 /// Represents a block of text that has been extruded.
 class ARKitText extends ARKitGeometry {
   ARKitText({
-    @required Vector3 position,
     @required this.text,
     @required this.extrusionDepth,
-    Vector3 scale,
-    Vector4 rotation,
     List<ARKitMaterial> materials,
-    String name,
   }) : super(
-          position,
           materials: materials,
-          scale: scale,
-          rotation: rotation,
-          name: name,
         );
 
   /// The text to be represented.