+++ lib/arkit_plugin.dart
@@ -7,3 +7,6 @@ export 'package:arkit_plugin/geometries/arkit_material.dart';
 export 'package:arkit_plugin/geometries/arkit_sphere.dart';
 export 'package:arkit_plugin/geometries/arkit_plane.dart';
 export 'package:arkit_plugin/geometries/arkit_text.dart';
+export 'package:arkit_plugin/physics/arkit_physics_body_type.dart';
+export 'package:arkit_plugin/physics/arkit_physics_body.dart';
+export 'package:arkit_plugin/physics/arkit_physics_shape.dart';