+++ lib/physics/arkit_physics_body_type.dart
@@ -1,0 +1,6 @@
+/// Physics Body type
+enum ARKitPhysicsBodyType {
+  staticType,
+  dynamicType,
+  kinematic,
+}