+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.3
 
 * Add debug options: world origin, feature points
+* ARPlaneAnchor added
 
 ## 0.0.2