+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.0.4
+
 ## 0.0.3
 
 * ARKitNode added